+++ providers/edge/src/airflow/providers/edge/example_dags/win_test.py
@@ -37,9 +37,9 @@ from airflow.exceptions import AirflowException, AirflowNotFoundException, Airfl
 from airflow.hooks.base import BaseHook
 from airflow.models import BaseOperator
 from airflow.models.dag import DAG
-from airflow.models.param import Param
 from airflow.models.variable import Variable
 from airflow.operators.empty import EmptyOperator
+from airflow.sdk.definitions.param import Param
 from airflow.utils.operator_helpers import context_to_airflow_vars
 from airflow.utils.trigger_rule import TriggerRule
 from airflow.utils.types import ArgNotSet
