+++ tests/serialization/test_dag_serialization.py
@@ -63,13 +63,13 @@ from airflow.models.dag import DAG
 from airflow.models.dagbag import DagBag
 from airflow.models.expandinput import EXPAND_INPUT_EMPTY
 from airflow.models.mappedoperator import MappedOperator
-from airflow.models.param import Param, ParamsDict
 from airflow.models.xcom import XCom
 from airflow.operators.empty import EmptyOperator
 from airflow.providers.cncf.kubernetes.pod_generator import PodGenerator
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.providers.standard.sensors.bash import BashSensor
 from airflow.sdk.definitions.asset import Asset
+from airflow.sdk.definitions.param import Param, ParamsDict
 from airflow.security import permissions
 from airflow.serialization.enums import Encoding
 from airflow.serialization.json_schema import load_dag_schema_dict
@@ -2181,7 +2181,7 @@ class TestStringifiedDAGs:
                 "timezone": "UTC",
                 "params": {
                     "my_param": {
-                        "__class": "airflow.models.param.Param",
+                        "__class": "airflow.sdk.definitions.param.Param",
                         "default": "str",
                     }
                 },
@@ -2204,7 +2204,7 @@ class TestStringifiedDAGs:
                 "fileloc": "/path/to/file.py",
                 "tasks": [],
                 "timezone": "UTC",
-                "params": [["str", {"__class": "airflow.models.param.Param", "default": "str"}]],
+                "params": [["str", {"__class": "airflow.sdk.definitions.param.Param", "default": "str"}]],
             },
         }
         SerializedDAG.validate_schema(serialized)
@@ -2228,7 +2228,7 @@ class TestStringifiedDAGs:
                             "default": "a string value",
                             "description": "hello",
                             "schema": {"__var": {"type": "string"}, "__type": "dict"},
-                            "__class": "airflow.models.param.Param",
+                            "__class": "airflow.sdk.definitions.param.Param",
                         },
                     ]
                 ],
